requirements.txt
python-telegram-bot==13.7
django-environ
django-apscheduler
telethon
django-simpleui
pymysql
django-q
requests==2.31.0
python-dotenv
cryptography
pylibmc
django==4.2
django-db-connection-pool
gunicorn
Pillow
djangorestframework==3.15.2
djangorestframework-simplejwt
django-storages